requirements.txt
Flask
Flask-SQLAlchemy
Flask-PyMongo
psycopg2-binary
python-dotenv
pytest
pytest-flask